Removed dependency on usbhid
In 2.6.28, the only additional configuration needed is for the new
macbooks using the wellspring3 device. The configuration is already
upstream, in a stalled commit ac26fca3 which is in hid-core, not
usbhid. Since jaunty comes with hid built in, there is no need or
possibility to use dkms for the configuration, but one must resort
to other methods.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>